move FileDownloader to downloaders module
