+++ conans/client/rest/rest_client_v2.py
@@ -8,7 +8,7 @@ from conans.client.rest.client_routes import ClientV2Router
 from conans.client.rest.download_cache import CachedFileDownloader
 from conans.client.rest.file_uploader import FileUploader
 from conans.client.rest.rest_client_common import RestCommonMethods, get_exception_from_error
-from conans.client.rest.file_downloader import FileDownloader
+from conans.client.downloaders.file_downloader import FileDownloader
 from conans.errors import ConanException, NotFoundException, PackageNotFoundException, \
     RecipeNotFoundException, AuthenticationException, ForbiddenException
 from conans.model.info import ConanInfo
