+++ conans/test/unittests/client/rest/downloader_test.py
@@ -2,7 +2,7 @@ import re
 import tempfile
 import unittest
 
-from conans.client.rest.file_downloader import FileDownloader
+from conans.client.downloaders.file_downloader import FileDownloader
 from conans.errors import ConanException
 from conans.test.utils.mocks import TestBufferConanOutput
 from conans.util.files import load
